engine/requirements.txt
flask
flask-cors
pandas
peewee
numpy
fastapi
uvicorn
requests
httpx
pydantic
python-jose[cryptography]
passlib[argon2]
python-multipart